requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.8.0
certifi==2019.9.11
cffi==1.12.3
cryptography==2.7
feedparser==5.2.1
future==0.17.1
html2markdown==0.1.7
pycparser==2.19
python-telegram-bot==12.0.0
six==1.12.0
soupsieve==1.9.3
tornado==6.0.3
